move analyse_Data route to agent_routs
